✨ feat: added new dependencies and updated configurations
Added `http-interop/http-factory-guzzle` and `meilisearch/meilisearch-php` to `composer.json` and `composer.lock`. Updated `config/scout-keys.php` to cast `SCOUT_KEY_EXPIRATION` as an int. Updated `search_keys` table migration to include new fields `engine` and `engine_key_id`.

+++ config/scout-keys.php
@@ -11,6 +11,6 @@ return [
         /**
          * The number of minutes a search key should be valid for.
          */
-        'lifetime' => env('SCOUT_KEY_EXPIRATION', 60 * 12),
+        'lifetime' => (int) env('SCOUT_KEY_EXPIRATION', 60 * 12),
     ],
 ];
